time print

+++ philo_one/putnbr.c
@@ -5,7 +5,7 @@ void	ft_putchar_fd(char c, int fd)
 	write(fd, &c, 1);
 }
 
-void	ft_putnbr_fd(int n, int fd)
+void	ft_putnbr_fd(long int n, int fd)
 {
 	if (n == -2147483648)
 	{
